+++ figforge/__init__.py
@@ -1,3 +1,3 @@
 """FigForge scientific figure assembly application."""
 
-__version__ = "0.13.0b1"
+__version__ = "0.14.0b1"